GluonTS Environment Connection Utilities

These functions are used for testing and establishing a python environment connection with modeltime.gluonts.

Usage

is_gluonts_activated()

activate_gluonts()

get_python_env()

check_gluonts_dependencies()

check_pytorch_dependencies()

detect_default_gluonts_env()

Details

is_gluonts_activated(): Determines if a GluonTS Environment has been activated during library(modeltime.gluonts).

If TRUE, then you should be good to go.

If FALSE, then a connection between modeltime.gluonts and your GluonTS Python has not been activated.

activate_gluonts(): Attempts to activate a connection between modeltime.gluonts and an associated GluonTS Python Environment using reticulate::use_condaenv(required = TRUE).

It first looks for the system environment variable, 'GLUONTS_PYTHON', for a path to the python executable

It next looks for a Conda Environment named 'r-gluonts' (this is what most users will have)

get_python_env(): Returns the configuration for the python environment that is being discovered using reticulate::py_discover_config().

check_gluonts_dependencies(): Checks whether GluonTS required python dependencies are present in the currently activated Python Environment.

detect_default_gluonts_env(): Detects if an 'r-gluonts' python environment is available.

Returns a tibble containing the

Returns NULL if an 'r-gluonts' environment is not detected

See also

Examples

# \donttest{
# Returns TRUE if GluonTS connection established on package load
is_gluonts_activated()
#> [1] TRUE

#

# }