GluonTS GP Forecaster Modeling Function (Bridge)

GluonTS GP Forecaster Modeling Function (Bridge)

gp_forecaster_fit_impl(
  x,
  y,
  freq,
  prediction_length,
  id,
  epochs = 5,
  num_batches_per_epoch = 50,
  learning_rate = 0.001,
  learning_rate_decay_factor = 0.5,
  patience = 10,
  minimum_learning_rate = 5e-05,
  clip_gradient = 10,
  weight_decay = 1e-08,
  init = "xavier",
  ctx = NULL,
  hybridize = TRUE,
  context_length = NULL,
  cardinality = 1,
  kernel_output = NULL,
  params_scaling = TRUE,
  max_iter_jitter = 10,
  jitter_method = "iter",
  time_features = NULL,
  num_parallel_samples = 100,
  batch_size = 32,
  scale_by_id = FALSE,
  ...
)

Arguments

x
A dataframe of xreg (exogenous regressors)

y
A numeric vector of values to fit

freq
A pandas timeseries frequency such as "5min" for 5-minutes or "D" for daily. Refer to Pandas Offset Aliases.

prediction_length
Numeric value indicating the length of the prediction horizon

id
A quoted column name that tracks the GluonTS FieldName "item_id"

epochs
Number of epochs that the network will train (default: 5).

num_batches_per_epoch
Number of batches at each epoch (default: 50).

learning_rate
Initial learning rate (default: 10-3 ).

learning_rate_decay_factor
Factor (between 0 and 1) by which to decrease the learning rate (default: 0.5).

patience
The patience to observe before reducing the learning rate, nonnegative integer (default: 10).

minimum_learning_rate
Lower bound for the learning rate (default: 5x10-5 ).

clip_gradient
Maximum value of gradient. The gradient is clipped if it is too large (default: 10).

weight_decay
The weight decay (or L2 regularization) coefficient. Modifies objective by adding a penalty for having large weights (default 10-8 ).

init
Initializer of the weights of the network (default: “xavier”).

ctx
The mxnet CPU/GPU context. Refer to using CPU/GPU in the mxnet documentation. (default: NULL, uses CPU)

hybridize
Increases efficiency by using symbolic programming. (default: TRUE)

context_length
Number of steps to unroll the RNN for before computing predictions (default: NULL, in which case context_length = prediction_length)

cardinality
Number of time series. Default: 1

kernel_output
KernelOutput instance to determine which kernel subclass to be instantiated (default: RBFKernelOutput()).

params_scaling
Determines whether or not to scale the model parameters (default: True).

max_iter_jitter
Maximum number of iterations for jitter to iteratively make the matrix positive definite (default: 10).

jitter_method
Iteratively jitter method or use eigenvalue decomposition depending on problem size (default: "iter").

time_features
Time features to use as inputs of the RNN (default: None, in which case these are automatically determined based on freq)

num_parallel_samples
Number of evaluation samples per time series to increase parallelism during inference. This is a model optimization that does not affect the accuracy (default: 100)

batch_size
Number of examples in each batch (default: 32).

scale_by_id
Scales numeric data by id group using mean = 0, standard deviation = 1 transformation. (default: FALSE)

...
Additional parameters passed to gluonts.model.gp_forecaster.GP ForecasterEstimator()